requirements.txt
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
click==6.7
command-not-found==0.3
configparser==3.5.0
cryptography==2.1.4
cupshelpers==1.0
DateTime==4.3
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dnspython==1.16.0
dominate==2.4.0
entrypoints==0.3
env==0.1.0
environs==7.0.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Cors==3.0.8
Flask-DotEnv==0.1.2
Flask-HTTPAuth==3.2.3
Flask-JWT==0.3.2
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
Flask-Testing==0.7.1
Flask-WTF==0.14.2
futures==3.1.1
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.15.2
Jinja2==2.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kafka==1.3.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
marshmallow==3.3.0
mistune==0.8.4
mongoengine==0.19.1
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
notebook==6.0.2
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.2.1
pickleshare==0.7.5
pika==1.1.0
Pillow==5.1.0
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.0.0
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.5.2
pygobject==3.26.1
PyJWT==1.4.2
pymacaroons==0.13.0
pymongo==3.9.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.15.6
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.10.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.1
qtconsole==4.6.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.10.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
visitor==0.1.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.5.1
WTForms==2.2.1
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
